+++ template/dartway_starter_server/migrations/20260718065953622/migration.sql
@@ -1,7 +1,7 @@
 BEGIN;
 
 --
--- Class AppSetting as table app_setting
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "app_setting" (
     "id" bigserial PRIMARY KEY,
@@ -13,7 +13,7 @@ CREATE TABLE "app_setting" (
 CREATE UNIQUE INDEX "app_setting_key_unique_idx" ON "app_setting" USING btree ("settingKey");
 
 --
--- Class UserProfile as table user_profile
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "user_profile" (
     "id" bigserial PRIMARY KEY,
@@ -25,12 +25,12 @@ CREATE TABLE "user_profile" (
     "lastName" text,
     "imageUrl" text,
     "gender" text,
-    "role" text NOT NULL DEFAULT 'client'::text,
+    "role" text NOT NULL DEFAULT 'user'::text,
     "testVerificationCode" text
 );
 
 --
--- Class CloudStorageEntry as table serverpod_cloud_storage
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_cloud_storage" (
     "id" bigserial PRIMARY KEY,
@@ -47,7 +47,7 @@ CREATE UNIQUE INDEX "serverpod_cloud_storage_path_idx" ON "serverpod_cloud_stora
 CREATE INDEX "serverpod_cloud_storage_expiration" ON "serverpod_cloud_storage" USING btree ("expiration");
 
 --
--- Class CloudStorageDirectUploadEntry as table serverpod_cloud_storage_direct_upload
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_cloud_storage_direct_upload" (
     "id" bigserial PRIMARY KEY,
@@ -61,7 +61,7 @@ CREATE TABLE "serverpod_cloud_storage_direct_upload" (
 CREATE UNIQUE INDEX "serverpod_cloud_storage_direct_upload_storage_path" ON "serverpod_cloud_storage_direct_upload" USING btree ("storageId", "path");
 
 --
--- Class FutureCallEntry as table serverpod_future_call
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_future_call" (
     "id" bigserial PRIMARY KEY,
@@ -78,7 +78,7 @@ CREATE INDEX "serverpod_future_call_serverId_idx" ON "serverpod_future_call" USI
 CREATE INDEX "serverpod_future_call_identifier_idx" ON "serverpod_future_call" USING btree ("identifier");
 
 --
--- Class ServerHealthConnectionInfo as table serverpod_health_connection_info
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_health_connection_info" (
     "id" bigserial PRIMARY KEY,
@@ -94,7 +94,7 @@ CREATE TABLE "serverpod_health_connection_info" (
 CREATE UNIQUE INDEX "serverpod_health_connection_info_timestamp_idx" ON "serverpod_health_connection_info" USING btree ("timestamp", "serverId", "granularity");
 
 --
--- Class ServerHealthMetric as table serverpod_health_metric
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_health_metric" (
     "id" bigserial PRIMARY KEY,
@@ -110,7 +110,7 @@ CREATE TABLE "serverpod_health_metric" (
 CREATE UNIQUE INDEX "serverpod_health_metric_timestamp_idx" ON "serverpod_health_metric" USING btree ("timestamp", "serverId", "name", "granularity");
 
 --
--- Class LogEntry as table serverpod_log
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_log" (
     "id" bigserial PRIMARY KEY,
@@ -130,7 +130,7 @@ CREATE TABLE "serverpod_log" (
 CREATE INDEX "serverpod_log_sessionLogId_idx" ON "serverpod_log" USING btree ("sessionLogId");
 
 --
--- Class MessageLogEntry as table serverpod_message_log
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_message_log" (
     "id" bigserial PRIMARY KEY,
@@ -147,7 +147,7 @@ CREATE TABLE "serverpod_message_log" (
 );
 
 --
--- Class MethodInfo as table serverpod_method
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_method" (
     "id" bigserial PRIMARY KEY,
@@ -159,7 +159,7 @@ CREATE TABLE "serverpod_method" (
 CREATE UNIQUE INDEX "serverpod_method_endpoint_method_idx" ON "serverpod_method" USING btree ("endpoint", "method");
 
 --
--- Class DatabaseMigrationVersion as table serverpod_migrations
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_migrations" (
     "id" bigserial PRIMARY KEY,
@@ -172,7 +172,7 @@ CREATE TABLE "serverpod_migrations" (
 CREATE UNIQUE INDEX "serverpod_migrations_ids" ON "serverpod_migrations" USING btree ("module");
 
 --
--- Class QueryLogEntry as table serverpod_query_log
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_query_log" (
     "id" bigserial PRIMARY KEY,
@@ -192,7 +192,7 @@ CREATE TABLE "serverpod_query_log" (
 CREATE INDEX "serverpod_query_log_sessionLogId_idx" ON "serverpod_query_log" USING btree ("sessionLogId");
 
 --
--- Class ReadWriteTestEntry as table serverpod_readwrite_test
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_readwrite_test" (
     "id" bigserial PRIMARY KEY,
@@ -200,7 +200,7 @@ CREATE TABLE "serverpod_readwrite_test" (
 );
 
 --
--- Class RuntimeSettings as table serverpod_runtime_settings
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_runtime_settings" (
     "id" bigserial PRIMARY KEY,
@@ -211,7 +211,7 @@ CREATE TABLE "serverpod_runtime_settings" (
 );
 
 --
--- Class SessionLogEntry as table serverpod_session_log
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "serverpod_session_log" (
     "id" bigserial PRIMARY KEY,
@@ -238,7 +238,7 @@ CREATE INDEX "serverpod_session_log_touched_idx" ON "serverpod_session_log" USIN
 CREATE INDEX "serverpod_session_log_isopen_idx" ON "serverpod_session_log" USING btree ("isOpen");
 
 --
--- Class DwAppNotification as table dw_app_notification
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "dw_app_notification" (
     "id" bigserial PRIMARY KEY,
@@ -252,7 +252,7 @@ CREATE TABLE "dw_app_notification" (
 );
 
 --
--- Class DwAuthKey as table dw_auth_key
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "dw_auth_key" (
     "id" bigserial PRIMARY KEY,
@@ -264,7 +264,7 @@ CREATE TABLE "dw_auth_key" (
 CREATE INDEX "dw_auth_key_userId_idx" ON "dw_auth_key" USING btree ("userId");
 
 --
--- Class DwAuthRequest as table dw_auth_request
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "dw_auth_request" (
     "id" bigserial PRIMARY KEY,
@@ -281,7 +281,7 @@ CREATE TABLE "dw_auth_request" (
 );
 
 --
--- Class DwAuthVerification as table dw_auth_verification
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "dw_auth_verification" (
     "id" bigserial PRIMARY KEY,
@@ -292,7 +292,7 @@ CREATE TABLE "dw_auth_verification" (
 );
 
 --
--- Class DwCloudFile as table dw_cloud_file
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "dw_cloud_file" (
     "id" bigserial PRIMARY KEY,
@@ -307,7 +307,7 @@ CREATE TABLE "dw_cloud_file" (
 );
 
 --
--- Class DwUserPassword as table dw_user_password
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "dw_user_password" (
     "id" bigserial PRIMARY KEY,
@@ -321,7 +321,7 @@ CREATE TABLE "dw_user_password" (
 CREATE UNIQUE INDEX "dw_user_password_user_id_idx" ON "dw_user_password" USING btree ("userId");
 
 --
--- Class DwWebServerLog as table dw_web_server_log
+-- ACTION CREATE TABLE
 --
 CREATE TABLE "dw_web_server_log" (
     "id" bigserial PRIMARY KEY,
@@ -339,7 +339,7 @@ CREATE TABLE "dw_web_server_log" (
 );
 
 --
--- Foreign relations for "serverpod_log" table
+-- ACTION CREATE FOREIGN KEY
 --
 ALTER TABLE ONLY "serverpod_log"
     ADD CONSTRAINT "serverpod_log_fk_0"
@@ -349,7 +349,7 @@ ALTER TABLE ONLY "serverpod_log"
     ON UPDATE NO ACTION;
 
 --
--- Foreign relations for "serverpod_message_log" table
+-- ACTION CREATE FOREIGN KEY
 --
 ALTER TABLE ONLY "serverpod_message_log"
     ADD CONSTRAINT "serverpod_message_log_fk_0"
@@ -359,7 +359,7 @@ ALTER TABLE ONLY "serverpod_message_log"
     ON UPDATE NO ACTION;
 
 --
--- Foreign relations for "serverpod_query_log" table
+-- ACTION CREATE FOREIGN KEY
 --
 ALTER TABLE ONLY "serverpod_query_log"
     ADD CONSTRAINT "serverpod_query_log_fk_0"
@@ -369,7 +369,7 @@ ALTER TABLE ONLY "serverpod_query_log"
     ON UPDATE NO ACTION;
 
 --
--- Foreign relations for "dw_auth_verification" table
+-- ACTION CREATE FOREIGN KEY
 --
 ALTER TABLE ONLY "dw_auth_verification"
     ADD CONSTRAINT "dw_auth_verification_fk_0"
@@ -383,9 +383,9 @@ ALTER TABLE ONLY "dw_auth_verification"
 -- MIGRATION VERSION FOR dartway_starter
 --
 INSERT INTO "serverpod_migrations" ("module", "version", "timestamp")
-    VALUES ('dartway_starter', '20260714113459228', now())
+    VALUES ('dartway_starter', '20260718065953622', now())
     ON CONFLICT ("module")
-    DO UPDATE SET "version" = '20260714113459228', "timestamp" = now();
+    DO UPDATE SET "version" = '20260718065953622', "timestamp" = now();
 
 --
 -- MIGRATION VERSION FOR serverpod